internal: types modularization
